refactor: 設定画面コンポーネントの分割（S15）
character.tsx（1510行）/ voice.tsx（1367行）/ index.tsx（998行）を、
既存の modelProvider/ サブコンポーネント分割パターンに倣って
character/・voice/・shell/ 配下に分解。各設定ブロック（キャラクター
モデル・ポーズ・TTSエンジン別設定・検索/ヘッダー/フッター等）を
1コンポーネントに切り出し、元ファイルは組み立てのみを行う薄い
エントリポイントとして残した。

JSX・ロジック・DOM構造は一切変更していない純粋なコード移動。
previewツールで character/voice設定タブのスクロール全域・TTS
エンジン切り替え・検索ハイライトの見た目を分割前後で目視比較し、
差異がないことを確認。全1821テスト・lint・tsc・build確認済み。
