+++ src/components/settings/character/Live2DSettingsForm.tsx
@@ -1,0 +1,314 @@
+import { logger } from '@/lib/logger'
+import { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+
+import settingsStore from '@/features/stores/settings'
+
+const emotionFields = [
+  {
+    key: 'neutralEmotions',
+    label: 'Neutral Emotions',
+    defaultValue: ['Neutral'],
+  },
+  {
+    key: 'happyEmotions',
+    label: 'Happy Emotions',
+    defaultValue: ['Happy,Happy2'],
+  },
+  {
+    key: 'sadEmotions',
+    label: 'Sad Emotions',
+    defaultValue: ['Sad,Sad2,Troubled'],
+  },
+  {
+    key: 'angryEmotions',
+    label: 'Angry Emotions',
+    defaultValue: ['Angry,Focus'],
+  },
+  {
+    key: 'relaxedEmotions',
+    label: 'Relaxed Emotions',
+    defaultValue: ['Relaxed'],
+  },
+  {
+    key: 'surprisedEmotions',
+    label: 'Surprised Emotions',
+    defaultValue: ['Surprised'],
+  },
+] as const
+
+const motionFields = [
+  { key: 'idleMotionGroup', label: 'Idle Motion Group', defaultValue: 'Idle' },
+  {
+    key: 'neutralMotionGroup',
+    label: 'Neutral Motion Group',
+    defaultValue: 'Neutral',
+  },
+  {
+    key: 'happyMotionGroup',
+    label: 'Happy Motion Group',
+    defaultValue: 'Happy',
+  },
+  { key: 'sadMotionGroup', label: 'Sad Motion Group', defaultValue: 'Sad' },
+  {
+    key: 'angryMotionGroup',
+    label: 'Angry Motion Group',
+    defaultValue: 'Angry',
+  },
+  {
+    key: 'relaxedMotionGroup',
+    label: 'Relaxed Motion Group',
+    defaultValue: 'Relaxed',
+  },
+  {
+    key: 'surprisedMotionGroup',
+    label: 'Surprised Motion Group',
+    defaultValue: 'Surprised',
+  },
+] as const
+
+interface Live2DModel {
+  path: string
+  name: string
+  expressions: string[]
+  motions: string[]
+}
+
+type EmotionFieldKey = (typeof emotionFields)[number]['key']
+
+export const Live2DSettingsForm = () => {
+  const store = settingsStore()
+  const { t } = useTranslation()
+  const [currentModel, setCurrentModel] = useState<Live2DModel | null>(null)
+  const [openDropdown, setOpenDropdown] = useState<EmotionFieldKey | null>(null)
+
+  useEffect(() => {
+    // 現在選択されているLive2Dモデルの情報を取得
+    const fetchCurrentModel = async () => {
+      try {
+        const response = await fetch('/api/get-live2d-list')
+        const models: Live2DModel[] = await response.json()
+        const selected = models.find(
+          (model) => model.path === store.selectedLive2DPath
+        )
+        setCurrentModel(selected || null)
+      } catch (error) {
+        logger.error('Error fetching Live2D model info:', error)
+      }
+    }
+
+    if (store.selectedLive2DPath) {
+      fetchCurrentModel()
+    }
+  }, [store.selectedLive2DPath])
+
+  // コンポーネントマウント時にデフォルト値を設定
+  useEffect(() => {
+    const updates: Record<string, any> = {}
+
+    emotionFields.forEach((field) => {
+      if (!store[field.key] || store[field.key].length === 0) {
+        updates[field.key] = field.defaultValue
+      }
+    })
+
+    motionFields.forEach((field) => {
+      if (!store[field.key] || store[field.key] === '') {
+        updates[field.key] = field.defaultValue
+      }
+    })
+
+    if (Object.keys(updates).length > 0) {
+      settingsStore.setState(updates)
+    }
+  }, [])
+
+  const handleEmotionChange = (
+    key: EmotionFieldKey,
+    expression: string,
+    checked: boolean
+  ) => {
+    const currentValues = store[key]
+    const newValues = checked
+      ? [...currentValues, expression]
+      : currentValues.filter((value) => value !== expression)
+
+    settingsStore.setState({
+      [key]: newValues,
+    })
+  }
+
+  const handleMotionChange = (key: string, value: string) => {
+    settingsStore.setState({
+      [key]: value,
+    })
+  }
+
+  if (!currentModel) {
+    return (
+      <div className="flex items-center justify-center h-32 text-gray-500">
+        {t('Live2D.LoadingModel')}
+      </div>
+    )
+  }
+
+  return (
+    <div className="space-y-8">
+      <div className="mb-6">
+        <div className="mb-4 text-xl font-bold">{t('Live2D.Emotions')}</div>
+        <div className="my-2 text-sm whitespace-pre-wrap">
+          {t('Live2D.EmotionInfo')}
+        </div>
+        <div className="space-y-4 text-sm">
+          {emotionFields.map((field) => (
+            <div key={field.key}>
+              <label className="block mb-2 font-bold">
+                {t(`Live2D.${field.key}`)}
+              </label>
+              <div className="relative">
+                <button
+                  type="button"
+                  className="w-full px-2 py-2 bg-white hover:bg-white-hover rounded-lg text-left flex items-center justify-between"
+                  onClick={() =>
+                    setOpenDropdown(
+                      openDropdown === field.key ? null : field.key
+                    )
+                  }
+                >
+                  <div className="flex flex-wrap gap-1">
+                    {store[field.key].length > 0 ? (
+                      store[field.key].map((expression) => (
+                        <span
+                          key={expression}
+                          className="inline-flex items-center px-2 py-1 bg-gray-100 rounded-lg mr-1"
+                        >
+                          {expression}
+                          <button
+                            type="button"
+                            className="ml-4 text-gray-500 hover:text-gray-700"
+                            onClick={(e) => {
+                              e.stopPropagation()
+                              handleEmotionChange(field.key, expression, false)
+                            }}
+                          >
+                            <svg
+                              className="h-4 w-4"
+                              fill="none"
+                              viewBox="0 0 24 24"
+                              stroke="currentColor"
+                            >
+                              <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={2}
+                                d="M6 18L18 6M6 6l12 12"
+                              />
+                            </svg>
+                          </button>
+                        </span>
+                      ))
+                    ) : (
+                      <span className="">{t('Live2D.SelectEmotions')}</span>
+                    )}
+                  </div>
+                  <svg
+                    className={`h-4 w-4  transition-transform ${
+                      openDropdown === field.key ? 'rotate-180' : ''
+                    }`}
+                    viewBox="0 0 16 16"
+                    fill="none"
+                    stroke="currentColor"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={1.5}
+                      d="M2 5l6 6 6-6"
+                    />
+                  </svg>
+                </button>
+                {openDropdown === field.key && (
+                  <div className="absolute z-10 w-full mt-4 max-h-[200px] overflow-y-auto bg-white rounded-lg shadow-lg border-gray-200 divide-y divide-gray-200">
+                    {currentModel.expressions.map((expression) => (
+                      <label
+                        key={expression}
+                        className="flex items-center px-4 py-2 hover:bg-white-hover cursor-pointer"
+                      >
+                        <input
+                          type="checkbox"
+                          className="h-4 w-4 rounded border-gray-300 text-primary focus:ring-primary mr-2"
+                          checked={store[field.key].includes(expression)}
+                          onChange={(e) =>
+                            handleEmotionChange(
+                              field.key,
+                              expression,
+                              e.target.checked
+                            )
+                          }
+                        />
+                        <span className="">{expression}</span>
+                      </label>
+                    ))}
+                  </div>
+                )}
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+
+      <div className="">
+        <div className="mb-4 text-xl font-bold">{t('Live2D.MotionGroups')}</div>
+        <div className="my-2 text-sm whitespace-pre-wrap">
+          {t('Live2D.MotionGroupsInfo')}
+        </div>
+        <div className="space-y-4">
+          {motionFields.map((field) => (
+            <div key={field.key}>
+              <label className="block mb-2 font-bold">
+                {t(`Live2D.${field.key}`)}
+              </label>
+              <div className="relative">
+                <select
+                  className="w-full px-4 py-2 bg-white hover:bg-white-hover rounded-lg appearance-none cursor-pointer"
+                  value={store[field.key]}
+                  onChange={(e) =>
+                    handleMotionChange(field.key, e.target.value)
+                  }
+                >
+                  <option value="" className="">
+                    {t('Live2D.SelectMotionGroup')}
+                  </option>
+                  {currentModel.motions.map((motion) => (
+                    <option
+                      key={motion}
+                      value={motion}
+                      className="py-4 px-8 hover:bg-primary hover:text-theme"
+                    >
+                      {motion}
+                    </option>
+                  ))}
+                </select>
+                <div className="absolute inset-y-0 right-4 flex items-center pointer-events-none">
+                  <svg
+                    className="h-4 w-4 "
+                    viewBox="0 0 16 16"
+                    fill="none"
+                    stroke="currentColor"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={1.5}
+                      d="M2 5l6 6 6-6"
+                    />
+                  </svg>
+                </div>
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  )
+}
