+++ src/components/settings/shell/Footer.tsx
@@ -1,0 +1,7 @@
+export const Footer = () => {
+  return (
+    <footer className="theme-surface-contrast shrink-0 border-t border-primary/20 py-1 text-center font-Montserrat text-xs">
+      powered by ChatVRM from Pixiv / ver. 2.55.2
+    </footer>
+  )
+}
