+++ src/components/settings/shell/index.ts
@@ -1,0 +1,3 @@
+export { Header } from './Header'
+export { Main } from './Main'
+export { Footer } from './Footer'
